Adds Recaptcha V2 support
I got the version 2 implementation from hgGeorg's fork (https://github.com/hgGeorg/node-recaptcha/blob/master/lib/recaptcha.js)
and inserted it here but keeping the lib backwards compatible so
it is possible to make a smooth transition, or even support both
versions in the same code.
